Zero Cascade and Manual New Game
Clicking on a "zero" (e.g. a spot with no touching mines) now reveals all adjacent cells for as long as there are touching zero cells.
The game now no longer starts a new game as soon as the player wins or loses.
Fixed the issue where the game would not count a win when all cells were revealed when the last cell was revealed via the "revealSurrounding()" command.